4 projects, GPLauncher linked, GPEditor unused.
New projects added to the solution for a cleaner solution.
GPLauncher settings has been linked to correctly to other projects.
We can run the executable if we use GPLauncher as the default startup project.
Basic classes has been added as a base of the architecture and to visualize how things will be next.